Add thruster particle effect for main character
A particle is a single white pixel that can be moved, scaled, rotated, and
drawn with a different color.
A ParticleEffect groups similar Particles and adds new ones to a List when
Spawn is called.
ThrusterParticleEffect and ExhaustParticle are subclasses specialized to
create the main character's jetpack effect.
Play with the parameters to get the desired effect.