Add Black Hole Particle Effect.
Tried creating BlackHoleParticle : Particle and BlackHole : ParticleEffect
to create a black-hole particle effect. It looks ok, but I don't think it
turned out as well as the jetpack. It may work better as an addition to
some black hole sprite animation rather than a stand-alone effect to
represent the entire black hole.