Add gravity to black hole.
Added class Gravity to simplify application of gravitational pull to
sprites. Black hole now has a Gravity property. PhysicalSprite has an
Update method that takes a Gravity as a parameter and applies the effects
of that gravity.
Gravity is to be improved by adding support for addition of Gravity
objects. This would simplify situations with multiple gravities on the
stage at one time.