Black hole eats nearby enemies.
The black hole will absorb nearby PhysicalSprites.
This changes a state in the PhysicalSprite (represented with the enum
LifeState) which triggers a spinning and fading animation and eventually
chages the state to Destroyed, at which point the sprite is removed from
the list.

TODO: Add list of absorbed sprites to blackhole so it can spit them out.